+++ packages/api-server/src/db/Regulations.ts
@@ -12,7 +12,7 @@ import { getRegulationLawChapters } from './LawChapter';
 import { db } from '../utils/sequelize';
 import { BindOrReplacements, QueryTypes } from 'sequelize';
 import promiseAll from '@hugsmidjan/qj/promiseAllObject';
-import { eliminateComments } from '@island.is/regulations-tools/utils';
+import { eliminateComments } from '@island.is/regulations-tools/textHelpers';
 
 export const PER_PAGE = 30;